+++ nutzboot-starter/nutzboot-starter-jdbc/src/main/java/org/nutz/boot/starter/jdbc/DataSourceStarter.java
@@ -1,8 +1,12 @@
 package org.nutz.boot.starter.jdbc;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 import java.util.Properties;
+import java.util.Random;
 
 import javax.sql.DataSource;
 
@@ -14,6 +18,8 @@ import org.nutz.ioc.loader.annotation.Inject;
 import org.nutz.ioc.loader.annotation.IocBean;
 import org.nutz.lang.Lang;
 import org.nutz.lang.Strings;
+import org.nutz.log.Log;
+import org.nutz.log.Logs;
 
 import com.alibaba.druid.pool.DruidDataSource;
 import com.alibaba.druid.pool.DruidDataSourceFactory;
@@ -23,50 +29,52 @@ import com.zaxxer.hikari.HikariDataSource;
 @IocBean
 public class DataSourceStarter {
 
-	protected static String PRE = "jdbc.";
-	@PropDoc(group = "jdbc", value = "连接池类型", possible = { "druid", "simple", "hikari" }, defaultValue = "druid")
-	public static final String PROP_TYPE = PRE + "type";
-	@PropDoc(group = "jdbc", value = "JDBC URL", need = true)
-	public static final String PROP_URL = PRE + "url";
-	@PropDoc(group = "jdbc", value = "数据库用户名")
-	public static final String PROP_USERNAME = PRE + "username";
-	@PropDoc(group = "jdbc", value = "数据库密码")
-	public static final String PROP_PASSWORD = PRE + "password";
+    private static final Log log = Logs.get();
 
-	@Inject
-	protected PropertiesProxy conf;
+    protected static String PRE = "jdbc.";
+    @PropDoc(group = "jdbc", value = "连接池类型", possible = {"druid", "simple", "hikari"}, defaultValue = "druid")
+    public static final String PROP_TYPE = PRE + "type";
+    @PropDoc(group = "jdbc", value = "JDBC URL", need = true)
+    public static final String PROP_URL = PRE + "url";
+    @PropDoc(group = "jdbc", value = "数据库用户名")
+    public static final String PROP_USERNAME = PRE + "username";
+    @PropDoc(group = "jdbc", value = "数据库密码")
+    public static final String PROP_PASSWORD = PRE + "password";
 
-	@Inject("refer:$ioc")
-	protected Ioc ioc;
+    @Inject
+    protected PropertiesProxy conf;
 
-	@IocBean
-	public DataSource getDataSource() throws Exception {
-		return createDataSource(ioc, conf, PRE);
-	}
+    @Inject("refer:$ioc")
+    protected Ioc ioc;
 
-	@IocBean(name = "druidDataSource", depose = "close")
-	public DataSource createDruidDataSource() throws Exception {
+    @IocBean
+    public DataSource getDataSource() throws Exception {
+        return createDataSource(ioc, conf, PRE);
+    }
+
+    @IocBean(name = "druidDataSource", depose = "close")
+    public DataSource createDruidDataSource() throws Exception {
         if (Strings.isBlank(conf.get(PROP_URL))) {
             throw new RuntimeException("need jdbc.url");
         }
-		return createDruidDataSource(conf, PRE);
-	}
-	
-	@IocBean(name = "hikariDataSource", depose = "close")
-	public DataSource createHikariCPDataSource() throws Exception {
+        return createDruidDataSource(conf, PRE);
+    }
+
+    @IocBean(name = "hikariDataSource", depose = "close")
+    public DataSource createHikariCPDataSource() throws Exception {
         if (Strings.isBlank(conf.get(PROP_URL))) {
             throw new RuntimeException("need jdbc.url");
         }
-		return createHikariCPDataSource(conf, PRE);
-	}
+        return createHikariCPDataSource(conf, PRE);
+    }
 
-	protected static boolean isDruid(PropertiesProxy conf) {
-		String type = conf.get(PROP_TYPE, "druid");
-		return "druid".equals(type) || "com.alibaba.druid.pool.DruidDataSource".equals(type);
-	}
-	
-	public static DataSource createDataSource(Ioc ioc, PropertiesProxy conf, String prefix) {
-	    switch (conf.get(prefix + "type", "druid")) {
+    protected static boolean isDruid(PropertiesProxy conf) {
+        String type = conf.get(PROP_TYPE, "druid");
+        return "druid".equals(type) || "com.alibaba.druid.pool.DruidDataSource".equals(type);
+    }
+
+    public static DataSource createDataSource(Ioc ioc, PropertiesProxy conf, String prefix) {
+        switch (conf.get(prefix + "type", "druid")) {
         case "simple":
         case "org.nutz.dao.impl.SimpleDataSource":
             SimpleDataSource simpleDataSource = new SimpleDataSource();
@@ -88,18 +96,18 @@ public class DataSourceStarter {
             break;
         }
         throw new RuntimeException("not supported jdbc.type=" + conf.get("jdbc.type"));
-	}
+    }
 
-	@SuppressWarnings({"rawtypes", "unchecked"})
+    @SuppressWarnings({"rawtypes", "unchecked"})
     public static DataSource createDruidDataSource(PropertiesProxy conf, String prefix) throws Exception {
         Map map = Lang.filter(new HashMap(conf.toMap()), prefix, null, null, null);
         DruidDataSource dataSource = (DruidDataSource) DruidDataSourceFactory.createDataSource(map);
         if (!conf.has(prefix + "filters"))
             dataSource.setFilters("stat");
         return dataSource;
-	}
+    }
 
-	public static DataSource createHikariCPDataSource(PropertiesProxy conf, String prefix) throws Exception {
+    public static DataSource createHikariCPDataSource(PropertiesProxy conf, String prefix) throws Exception {
         Properties properties = new Properties();
         for (String key : conf.keys()) {
             if (!key.startsWith(prefix) || key.equals(prefix + "type"))
@@ -107,11 +115,49 @@ public class DataSourceStarter {
             if (key.equals(prefix + "url")) {
                 if (!conf.has(prefix + "jdbcUrl"))
                     properties.put("jdbcUrl", conf.get(key));
-            }
-            else {
+            } else {
                 properties.put(key.substring(5), conf.get(key));
             }
         }
         return new HikariDataSource(new HikariConfig(properties));
-	}
+    }
+
+    public static DataSource getSlaveDataSource(Ioc ioc, PropertiesProxy conf) {
+        if (ioc.has("slaveDataSource")) {
+            return ioc.get(DataSource.class, "slaveDataSource");
+        } else {
+            // 看看有多少从数据库被定义了
+            List<DataSource> slaveDataSources = new ArrayList<>();
+            for (String key : conf.keys()) {
+                if (key.startsWith("jdbc.slave.") && key.endsWith(".url")) {
+                    String slaveName = key.substring("jdbc.slave.".length(), key.length() - ".url".length());
+                    log.debug("found Slave DataSource name=" + slaveName);
+                    slaveDataSources.add(DataSourceStarter.createDataSource(ioc, conf, "jdbc.slave." + slaveName + "."));
+                }
+            }
+            // 如果的确定义了从数据库集合
+            if (slaveDataSources.size() > 0) {
+                if (slaveDataSources.size() == 1) {
+                    // 单个? 那就直接set吧
+                    return slaveDataSources.get(0);
+                } else {
+                    // 多个从数据源,使用DynaDataSource进行随机挑选
+                    // TODO 更多更精细的挑选策略(轮训/随机/可用性...)
+                    return new DynaDataSource(new Iterator<DataSource>() {
+                        protected Random random = new Random(System.currentTimeMillis());
+                        protected DataSource[] ds = slaveDataSources.toArray(new DataSource[slaveDataSources.size()]);
+
+                        public DataSource next() {
+                            return ds[random.nextInt(slaveDataSources.size())];
+                        }
+
+                        public boolean hasNext() {
+                            return true;
+                        }
+                    });
+                }
+            }
+        }
+        return null;
+    }
 }
